refactor: modularize config and shared helpers
