+++ src/lib/types/models.ts
@@ -1,0 +1,54 @@
+export interface ModelsResponse {
+  data: Array<Model>
+  object: string
+}
+
+interface ModelLimits {
+  max_context_window_tokens?: number
+  max_output_tokens?: number
+  max_prompt_tokens?: number
+  max_inputs?: number
+  vision?: {
+    max_prompt_image_size?: number
+    max_prompt_images?: number
+    supported_media_types?: Array<string>
+  }
+}
+
+interface ModelSupports {
+  max_thinking_budget?: number
+  min_thinking_budget?: number
+  tool_calls?: boolean
+  parallel_tool_calls?: boolean
+  dimensions?: boolean
+  streaming?: boolean
+  structured_outputs?: boolean
+  vision?: boolean
+  adaptive_thinking?: boolean
+  reasoning_effort?: Array<string>
+}
+
+interface ModelCapabilities {
+  family: string
+  limits: ModelLimits
+  object: string
+  supports: ModelSupports
+  tokenizer: string
+  type: string
+}
+
+export interface Model {
+  capabilities: ModelCapabilities
+  id: string
+  model_picker_enabled: boolean
+  name: string
+  object: string
+  preview: boolean
+  vendor: string
+  version: string
+  policy?: {
+    state: string
+    terms: string
+  }
+  supported_endpoints?: Array<string>
+}
