+++ src/lib/utils.ts
@@ -1,15 +1,10 @@
 import type { Context } from "hono"
 
-import consola from "consola"
 import { createHash, randomUUID } from "node:crypto"
-import { networkInterfaces } from "node:os"
 
 import type { AnthropicMessagesPayload } from "~/lib/types/anthropic"
+import type { ResponsesStream } from "~/lib/types/responses"
 
-import { getModels as getCopilotModels } from "~/services/copilot/get-models"
-import { getVSCodeVersion } from "~/services/get-vscode-version"
-
-import { getVSCodeDeviceId } from "./deviceid"
 import { state } from "./state"
 
 export const sleep = (ms: number) =>
@@ -20,156 +15,12 @@ export const sleep = (ms: number) =>
 export const isNullish = (value: unknown): value is null | undefined =>
   value === null || value === undefined
 
-// Periodically refresh models so long-running daemons pick up new SKUs.
-const MODELS_REFRESH_BASE_MS = 30 * 60 * 1000
-let modelsRefreshTimer: ReturnType<typeof setTimeout> | null = null
+export const isAsyncIterable = <T>(value: unknown): value is AsyncIterable<T> =>
+  Boolean(value)
+  && typeof (value as AsyncIterable<T>)[Symbol.asyncIterator] === "function"
 
-export const stopModelsRefreshLoop = () => {
-  if (modelsRefreshTimer) {
-    clearTimeout(modelsRefreshTimer)
-    modelsRefreshTimer = null
-  }
-}
-
-type ModelsFetcher = typeof getCopilotModels
-
-const refreshModels = async (fetcher: ModelsFetcher) => {
-  const prevIds = new Set(state.models?.data.map((m) => m.id) ?? [])
-  const models = await fetcher()
-  state.models = {
-    ...models,
-    data: models.data.filter(
-      (model) =>
-        model.policy?.state !== "disabled"
-        && (model.model_picker_enabled
-          || model.capabilities.type === "embeddings"),
-    ),
-  }
-  const nextIds = state.models.data.map((m) => m.id)
-  const added = nextIds.filter((id) => !prevIds.has(id))
-  if (added.length > 0) {
-    consola.info(`Models refresh: ${added.length} new`)
-  } else {
-    consola.debug(`Models refresh: no changes (${nextIds.length} total)`)
-  }
-}
-
-const scheduleModelsRefresh = (fetcher: ModelsFetcher, intervalMs: number) => {
-  const jitter = Math.floor(Math.random() * (intervalMs / 6))
-  const delay = intervalMs + jitter
-  consola.debug(
-    `Scheduling next models refresh in ${Math.round(delay / 1000)} seconds`,
-  )
-
-  stopModelsRefreshLoop()
-  modelsRefreshTimer = setTimeout(async () => {
-    try {
-      await refreshModels(fetcher)
-    } catch (error) {
-      consola.warn("Failed to refresh models, keeping previous cache.", error)
-    } finally {
-      scheduleModelsRefresh(fetcher, intervalMs)
-    }
-  }, delay)
-}
-
-export async function cacheModels(
-  fetcher: ModelsFetcher = getCopilotModels,
-  intervalMs: number = MODELS_REFRESH_BASE_MS,
-): Promise<void> {
-  await refreshModels(fetcher)
-  scheduleModelsRefresh(fetcher, intervalMs)
-}
-
-export const cacheVSCodeVersion = async () => {
-  const response = await getVSCodeVersion()
-  state.vsCodeVersion = response
-
-  consola.info(`Using VSCode version: ${response}`)
-}
-
-const invalidMacAddresses = new Set([
-  "00:00:00:00:00:00",
-  "ff:ff:ff:ff:ff:ff",
-  "ac:de:48:00:11:22",
-])
-
-function validateMacAddress(candidate: string): boolean {
-  const tempCandidate = candidate.replaceAll("-", ":").toLowerCase()
-  return !invalidMacAddresses.has(tempCandidate)
-}
-
-export function getMac(): string | null {
-  const ifaces = networkInterfaces()
-  // eslint-disable-next-line guard-for-in
-  for (const name in ifaces) {
-    const networkInterface = ifaces[name]
-    if (networkInterface) {
-      for (const { mac } of networkInterface) {
-        if (validateMacAddress(mac)) {
-          return mac
-        }
-      }
-    }
-  }
-  return null
-}
-
-export const cacheMacMachineId = () => {
-  const macAddress = getMac() ?? randomUUID()
-  state.macMachineId = createHash("sha256")
-    .update(macAddress, "utf8")
-    .digest("hex")
-  consola.debug(`Using machine ID: ${state.macMachineId}`)
-}
-
-export const cacheVsCodeDeviceId = async () => {
-  state.vsCodeDeviceId = await getVSCodeDeviceId()
-  consola.debug(`Using VSCode device ID: ${state.vsCodeDeviceId}`)
-}
-
-const SESSION_REFRESH_BASE_MS = 60 * 60 * 1000
-const SESSION_REFRESH_JITTER_MS = 20 * 60 * 1000
-let vsCodeSessionRefreshTimer: ReturnType<typeof setTimeout> | null = null
-
-const generateSessionId = () => {
-  state.vsCodeSessionId = randomUUID() + Date.now().toString()
-  consola.debug(`Generated VSCode session ID: ${state.vsCodeSessionId}`)
-}
-
-export const stopVsCodeSessionRefreshLoop = () => {
-  if (vsCodeSessionRefreshTimer) {
-    clearTimeout(vsCodeSessionRefreshTimer)
-    vsCodeSessionRefreshTimer = null
-  }
-}
-
-const scheduleSessionIdRefresh = () => {
-  const randomDelay = Math.floor(Math.random() * SESSION_REFRESH_JITTER_MS)
-  const delay = SESSION_REFRESH_BASE_MS + randomDelay
-  consola.debug(
-    `Scheduling next VSCode session ID refresh in ${Math.round(
-      delay / 1000,
-    )} seconds`,
-  )
-
-  stopVsCodeSessionRefreshLoop()
-  vsCodeSessionRefreshTimer = setTimeout(() => {
-    try {
-      generateSessionId()
-    } catch (error) {
-      consola.error("Failed to refresh session ID, rescheduling...", error)
-    } finally {
-      scheduleSessionIdRefresh()
-    }
-  }, delay)
-}
-
-export const cacheVsCodeSessionId = () => {
-  stopVsCodeSessionRefreshLoop()
-  generateSessionId()
-  scheduleSessionIdRefresh()
-}
+export const isResponsesStream = (value: unknown): value is ResponsesStream =>
+  isAsyncIterable<unknown>(value)
 
 interface PayloadMessage {
   role?: string
