+++ src/routes/messages/non-stream-translation.ts
@@ -1,5 +1,5 @@
 import type { ToolContentSupportType } from "~/lib/config"
-import type { Model } from "~/services/copilot/get-models"
+import type { Model } from "~/lib/types/models"
 
 import { state } from "~/lib/state"
 import {
