+++ src/routes/messages/preprocess.ts
@@ -1,4 +1,4 @@
-import type { Model } from "~/services/copilot/get-models"
+import type { Model } from "~/lib/types/models"
 
 import {
   COMPACT_AUTO_CONTINUE,
