+++ src/routes/provider/utils.ts
@@ -1,0 +1,27 @@
+import type { ModelConfig } from "~/lib/config"
+
+interface SamplingPayload {
+  temperature?: number | null
+  top_p?: number | null
+  top_k?: number | null
+}
+
+export const applyModelDefaults = (
+  payload: SamplingPayload,
+  modelConfig: ModelConfig | undefined,
+): void => {
+  payload.temperature ??= modelConfig?.temperature
+  payload.top_p ??= modelConfig?.topP
+  payload.top_k ??= modelConfig?.topK
+}
+
+export const applyMissingExtraBody = (
+  payload: Record<string, unknown>,
+  options: { extraBody: Record<string, unknown> | undefined },
+): void => {
+  for (const [key, value] of Object.entries(options.extraBody ?? {})) {
+    if (!Object.hasOwn(payload, key)) {
+      payload[key] = value
+    }
+  }
+}
