+++ src/start.ts
@@ -16,13 +16,13 @@ import { initProxyFromEnv } from "./lib/proxy"
 import { generateEnvScript } from "./lib/shell"
 import { state } from "./lib/state"
 import { logUser, setupCopilotToken } from "./lib/token"
+import { cacheModels } from "./services/copilot/models-cache"
 import {
   cacheMacMachineId,
-  cacheModels,
   cacheVSCodeVersion,
   cacheVsCodeSessionId,
   cacheVsCodeDeviceId,
-} from "./lib/utils"
+} from "./services/vscode-env"
 
 interface RunServerOptions {
   port: number
