+++ tests/models-family-selection.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, expect, test } from "bun:test"
 
-import type { Model } from "~/services/copilot/get-models"
+import type { Model } from "~/lib/types/models"
 
 import { getLatestModelForFamily } from "~/lib/models"
 import { state } from "~/lib/state"
